requirements/development.txt
-r base.txt

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-asyncio==0.23.6
factory-boy==3.3.0
freezegun==1.4.0

# Code quality
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0
django-stubs==4.2.7
djangorestframework-stubs==3.15.0

# Debug
django-debug-toolbar==4.3.0
ipython==8.23.0
django-extensions==3.2.3

# Docs
mkdocs==1.5.3
mkdocs-material==9.5.15
